Stuff
Too much stuff. Gotta get back on the wagon

+++ inc/classes/GameLogs.php
@@ -10,22 +10,20 @@ class GameLogs{
     unset($this->round);
   }
 
-  public function fetchRemoteFile($url, $file=null){
-    $get = $url."/".$file;
-    $curl = curl_init();
-    curl_setopt_array($curl, array(
-      CURLOPT_RETURNTRANSFER => TRUE,
-      CURLOPT_URL => $get,
-      CURLOPT_USERAGENT => "atlantaned.space log parser",
-      CURLOPT_SSL_VERIFYPEER => FALSE,
-      CURLOPT_SSL_VERIFYHOST => FALSE,
-      CURLOPT_FOLLOWLOCATION => TRUE,
-      CURLOPT_REFERER => "atlantaned.space",
-      CURLOPT_ENCODING => 'gzip',
-    ));
-    $logs = curl_exec($curl);
-    curl_close($curl);
-    return $logs;
+  public function getGameLogs(){
+    if(file_exists($this->round->logCache)){ //Load from cache
+      $this->round->logs = $this->getCachedLogs($this->round->logCache);
+      $this->round->fromCache = TRUE;
+    } else { //Load from remote
+      $this->reset();
+      $app = new app();
+      $this->round->logs = $app->getRemoteFile($this->round->logsURL,"/game.txt");
+      $this->round->attack = $app->getRemoteFile($this->round->logsURL,"/attack.txt");
+      $this->round->logs = $this->parseGameLog($this->round->logs,$this->round->attack);
+      $this->cacheParsedLogs($this->round->logs);
+      $this->round->fromCache = FALSE;
+    }
+    return $this->round;
   }
 
   public function resetLog(){
@@ -33,6 +31,11 @@ class GameLogs{
     if(1 >= $user->level){
       return returnError("You do not have permission to do this.");
     }
+    $this->reset();
+    return returnSuccess($this->round->logCache." deleted. Regenerating stats.");
+  }
+
+  public function reset(){
     $db = new database(true);
     $db->query("DELETE FROM explosion_log WHERE round = ?");
     $db->bind(1,$this->round->id);
@@ -48,12 +51,24 @@ class GameLogs{
     } catch (Exception $e) {
       return returnError("Database error: ".$e->getMessage());
     }
-    unlink($this->round->logCache);
-    return returnSuccess($this->round->logCache." deleted. Regenerating stats.");
+    if(defined('LOG_CACHE_MODE') && 'database' == LOG_CACHE_MODE){
+      $db->query("DELETE FROM round_logs WHERE round = ?");
+      $db->bind(1,$this->round->id);
+      try {
+        $db->execute();
+      } catch (Exception $e) {
+        return returnError("Database error: ".$e->getMessage());
+      }
+    } else {
+      if(file_exists($this->round->logCache)){
+        unlink($this->round->logCache);
+      }
+    }
   }
   
   public function getAvailableLogs(){
-    $files = $this->fetchRemoteFile($this->round->logsURL);
+    $app = new app();
+    $files = $app->getRemoteFile($this->round->logsURL);
     $files = strip_tags($files);
     $files = explode("\r\n",$files);
     $tmp = array();
@@ -63,15 +78,15 @@ class GameLogs{
         $tmp[] = $f[0].".gz";
       }
     }
-    $files = (object) $tmp;
-    return $files;
+    $this->round->availableLogs = (object) $tmp;
+    return $round;
   }
 
   public function parseGameLog($logs,$attack){
     $date = date('Y-m-d',strtotime($this->round->start_datetime));
     // var_dump($this->round->logsURL."/game.log.gz");
     //This method is ONLY for cleaning out useless lines from the logs
-
+    $logs = strip_tags($logs);
     //For the sake of transparency, this method is heavily documented
     //
     //-censored lines have no value in public logs, and are removed entirely
@@ -109,6 +124,7 @@ class GameLogs{
     //TODO: Randomize the #-# to stop people from (un)intentionally breaking
     //the parser
     $logs = preg_replace("/(\[)(\d{2}:\d{2}:\d{2})(])(GAME|ACCESS|SAY|OOC|ADMIN|EMOTE|WHISPER|PDA|CHAT|LAW|PRAY|COMMENT|VOTE|GHOST)(:\s)/",$date." $2#-#$4#-#", $logs);
+    $logs = preg_replace("/([ ][(])([\d]{1,3}[,][\d]{1,3}[,][\d]{1,3})([)])/","#-#$2", $logs);
 
     //UTF8 encode (hey look who was ahead of the curve)
     $logs = utf8_encode($logs);
@@ -119,20 +135,32 @@ class GameLogs{
     //Remove empty lines
     array_filter($logs);
 
-    //Further explode the log line 
     $explosions = array();
+    $antagLine = 0;
     $antags = array();
     $antagLines = array();
     $i = 0;
 
+    //Further explode the log line 
     foreach ($logs as &$log){
       $log = explode('#-#',$log);
+      if(isset($log[2]) && 'Antagonists at round end were...' == $log[2]){
+        $antagLine = $i;
+      }
+      if(isset($log[3])){
+        $coords = explode(',',$log[3]);
+        $log['coord_x'] = (int) $coords[0];
+        $log['coord_y'] = (int) $coords[1];
+        $log['coord_z'] = (int) $coords[2];
+        unset($log[3]);
+      } else {
+        $log['coord_x'] = null;
+        $log['coord_y'] = null;
+        $log['coord_z'] = null;
+      }
       if (isset($log[1]) && 'GAME' == $log[1]
         && strpos($log[2], 'Explosion with size (') !== FALSE){
         $explosions[] = (object) $this->parseExplosion($log);
-      }
-      if(isset($log[2]) && 'Antagonists at round end were...' == $log[2]){
-        $antagLine = $i;
       }
       $i++;
     }
@@ -153,11 +181,23 @@ class GameLogs{
       $attack = explode("---------------------\r\n",$attack);
       $attack = $attack[1];
       $attack = preg_replace("/(\[)(\d{2}:\d{2}:\d{2})(])(ATTACK)(:\s)/",$date." $2#-#$4#-#", $attack);
+      $attack = preg_replace("/([ ][(])([\d]{1,3}[,][\d]{1,3}[,][\d]{1,3})([)])/","#-#$2", $attack);
       $attack = explode("\r\n",$attack);
       array_filter($attack);
       array_pop($attack);
       foreach ($attack as &$a){
         $a = explode('#-#',$a);
+        if(isset($log[3])){
+          $coords = explode(',',$a[3]);
+          $a['coord_x'] = (int) $coords[0];
+          $a['coord_y'] = (int) $coords[1];
+          $a['coord_z'] = (int) $coords[2];
+          unset($a[3]);
+        } else {
+          $a['coord_x'] = null;
+          $a['coord_y'] = null;
+          $a['coord_z'] = null;
+        }
         $logs[] = $a;
       }
     }
@@ -171,29 +211,72 @@ class GameLogs{
     return $logs;
   }
 
-  public function getGameLogs(){
-    if(file_exists($this->round->logCache)){ //Load from cache
-      $this->round->logs = $this->getCachedLogs($this->round->logCache);
-      $this->round->fromCache = TRUE;
-    } else { //Load from remote
-      $this->round->logs = $this->fetchRemoteFile($this->round->logsURL,"game.txt.gz");
-      $this->round->attack = $this->fetchRemoteFile($this->round->logsURL,"attack.txt.gz");
-      $this->round->logs = $this->parseGameLog($this->round->logs,$this->round->attack);
-      $this->cacheParsedLogs($this->round->logs);
-      $this->round->fromCache = FALSE;
+  public function areRoundLogsInDB(){
+    $db = new database(TRUE);
+    $db->query("SELECT round FROM round_logs WHERE round = ? LIMIT 0,1");
+    $db->bind(1, $this->round->id);
+    try {
+      if($db->single()){
+        return true;
+      }
+    } catch (Exception $e) {
+      return returnError("Database error: ".$e->getMessage());
     }
-    return $this->round;
+    return false;
   }
 
   public function getCachedLogs($cache){
-    $logs = file_get_contents($cache);
-    return json_decode($logs);
+    if(defined('LOG_CACHE_MODE') && 'database' == LOG_CACHE_MODE){
+      $db = new database(TRUE);
+      $db->query("SELECT round_logs.timestamp as `0`,
+        round_logs.type as `1`,
+        round_logs.text as `2`,
+        round_logs.x,
+        round_logs.y,
+        round_logs.z
+        FROM round_logs WHERE round = ?");
+      $db->bind(1, $this->round->id);
+      try {
+        return $db->resultset(\PDO::FETCH_ASSOC);
+      } catch (Exception $e) {
+        var_dump("Database error: ".$e->getMessage());
+      }
+    } else {
+      $logs = file_get_contents($cache);
+      return json_decode($logs, TRUE);
+    }
   }
 
   public function cacheParsedLogs($round){
-    $logsavefile = fopen($this->round->logCache,"w+");
-    fwrite($logsavefile,json_encode($this->round->logs,JSON_UNESCAPED_UNICODE));
-    fclose($logsavefile);
+    if(defined('LOG_CACHE_MODE') && 'database' == LOG_CACHE_MODE){
+      $this->pushLogsToDB();
+    } else {
+      $logsavefile = fopen($this->round->logCache,"w+");
+      fwrite($logsavefile,json_encode($this->round->logs,JSON_FORCE_OBJECT |JSON_UNESCAPED_UNICODE | JSON_NUMERIC_CHECK | JSON_UNESCAPED_SLASHES));
+      fclose($logsavefile);
+    }
+  }
+
+  public function pushLogsToDB(){
+    $db = new database(TRUE);
+    $db->query("INSERT INTO round_logs 
+      (round, `timestamp`, type, `text`, x, y, z)
+      VALUES (?, ?, ?, ?, ?, ?, ?)");
+    // var_dump($this->round->logs);
+    foreach ($this->round->logs as $log){
+      $db->bind(1, $this->round->id);
+      $db->bind(2, $log[0]);
+      $db->bind(3, $log[1]);
+      $db->bind(4, $log[2]);
+      $db->bind(5, $log['x']);
+      $db->bind(6, $log['y']);
+      $db->bind(7, $log['z']);
+      try {
+        $db->execute();
+      } catch (Exception $e) {
+        var_dump("Database error: ".$e->getMessage());
+      }
+    }
   }
 
   public function extractDataFromLogs($explosions, $antags){
@@ -242,22 +325,19 @@ class GameLogs{
   }
 
   public function parseExplosion($log){
-    $e['round'] = $this->round->id;
+    $log[2] = str_replace('Explosion with size (', '', $log[2]);
+    $log[2] = explode(') in area ',$log[2]);
+    $e['area'] = $log[2][1];
     $e['time'] = $log[0];
-    $exp = str_replace('Explosion with size (', '', $log[2]);
-    $exp = explode(') in area ',$exp);
-    $exp[0] = explode(', ',$exp[0]);
-    $e['devestation'] = (int) $exp[0][0];
-    $e['heavy'] = (int) $exp[0][1];
-    $e['light'] = (int) $exp[0][2];
-    $e['flash'] = (int) $exp[0][3];
-    $exp[1] = explode(' (',$exp[1]);
-    $exp[1][1] = str_replace(')', '', $exp[1][1]);
-    $e['area'] = $exp[1][0];
-    $exp[1][1] = explode(',',$exp[1][1]);
-    $e['x'] = (int) $exp[1][1][0];
-    $e['y'] = (int) $exp[1][1][1];
-    $e['z'] = (int) $exp[1][1][2];
+    $e['round'] = $this->round->id;
+    $e['x'] = $log['coord_x'];
+    $e['y'] = $log['coord_y'];
+    $e['z'] = $log['coord_z'];
+    $log[2][0] = explode(', ',$log[2][0]);
+    $e['devestation'] = (int) $log[2][0][0];
+    $e['heavy'] =       (int) $log[2][0][0];
+    $e['light'] =       (int) $log[2][0][0];
+    $e['flash'] =       (int) $log[2][0][0];
     return $e;
   }
